+++ web/src/App.jsx
@@ -19,83 +19,36 @@
  * find current contact information at www.suse.com.
  */
 
-import React, { useEffect, useReducer } from "react";
+import React, { useEffect, useState } from "react";
+import { Outlet, Navigate } from "react-router-dom";
 import { useInstallerClient } from "./context/installer";
 
-import { PROBING, PROBED, INSTALLING, INSTALLED } from "./client/status";
-
-import DBusError from "./DBusError";
-import Overview from "./Overview";
-import ProbingProgress from "./ProbingProgress";
-import InstallationProgress from "./InstallationProgress";
-import InstallationFinished from "./InstallationFinished";
 import LoadingEnvironment from "./LoadingEnvironment";
 import Questions from "./Questions";
 
 import './assets/fonts.scss';
 import "./app.scss";
 
-const init = status => ({
-  loading: status === null,
-  probing: status === PROBING,
-  probed: status === PROBED,
-  installing: status === INSTALLING,
-  finished: status === INSTALLED,
-  dbusError: null
-});
-
-const reducer = (state, action) => {
-  switch (action.type) {
-    case "CHANGE_STATUS": {
-      return init(action.payload.status);
-    }
-    case "SET_DBUS_ERROR": {
-      return { ...state, dbusError: action.payload.error };
-    }
-    default: {
-      throw new Error(`Unsupported action type: ${action.type}`);
-    }
-  }
-};
-
-const renderMainContent = (state) => {
-  if (state.dbusError) return <DBusError />;
-  if (state.loading) return <LoadingEnvironment />;
-  if (state.probing) return <ProbingProgress />;
-  if (state.installing) return <InstallationProgress />;
-  if (state.finished) return <InstallationFinished />;
-
-  return <Overview />;
-};
-
+// TODO: add documentation
 function App() {
+  const [product, setProduct] = useState(null);
   const client = useInstallerClient();
-  const [state, dispatch] = useReducer(reducer, null, init);
 
   useEffect(() => {
-    client.manager.getStatus()
-      .then(status => dispatch({ type: "CHANGE_STATUS", payload: { status } }))
-      .catch(error => dispatch({ type: "SET_DBUS_ERROR", payload: { error } }));
-  }, [client.manager]);
+    return client.software.getSelectedProduct()
+      .then(setProduct)
+      .catch(() => setProduct(undefined));
+  }, [client.software]);
 
-  useEffect(() => {
-    return client.manager.onChange(changes => {
-      if ("Status" in changes) {
-        dispatch({ type: "CHANGE_STATUS", payload: { status: changes.Status } });
-      }
-    });
-  }, [client.manager]);
+  // FIXME: improve this
+  if (product === null) return <LoadingEnvironment text="Checking products..." />;
 
-  useEffect(() => {
-    return client.monitor.onDisconnect(() => {
-      dispatch({ type: "SET_DBUS_ERROR", payload: { error: "Connection lost" } });
-    });
-  }, [client.monitor]);
+  if (!product) return <Navigate to="products" replace={true} />;
 
   return (
     <>
       <Questions />
-      { renderMainContent(state) }
+      <Outlet />
     </>
   );
 }
